# Move error string creation behind function call to avoid allocations