@juliohm 2025-07-31
We should probably add a simple benchmark for the `HalfEdgeTopology` given that it is heavily optimized now.