@halleysfifthinc 2025-08-01
Ah, sorry! 
> We should probably add a simple benchmark for the HalfEdgeTopology given that it is heavily optimized now.

I hadn't realized/understood this to be an immediate request.